+++ src/lightning_habana/fabric/strategies/parallel.py
@@ -27,7 +27,6 @@ if module_available("lightning"):
     from lightning.fabric.plugins.collectives.torch_collective import default_pg_timeout
     from lightning.fabric.plugins.environments.cluster_environment import ClusterEnvironment
     from lightning.fabric.plugins.io.torch_io import TorchCheckpointIO
-    from lightning.fabric.plugins.precision import Precision
     from lightning.fabric.strategies.ddp import DDPStrategy
     from lightning.fabric.utilities.types import Optimizable
 elif module_available("pytorch_lightning"):
@@ -36,7 +35,6 @@ elif module_available("pytorch_lightning"):
     from lightning_fabric.plugins.collectives.torch_collective import default_pg_timeout
     from lightning_fabric.plugins.environments.cluster_environment import ClusterEnvironment
     from lightning_fabric.plugins.io.torch_io import TorchCheckpointIO
-    from lightning_fabric.plugins.precision import Precision
     from lightning_fabric.strategies.ddp import DDPStrategy
     from lightning_fabric.utilities.types import Optimizable
 else:
@@ -47,6 +45,7 @@ from torch.nn import Module
 
 from lightning_habana import _HPU_AVAILABLE
 from lightning_habana.fabric.accelerator import HPUAccelerator
+from lightning_habana.fabric.plugins.precision import HPUPrecision
 from lightning_habana.utils.imports import _HABANA_FRAMEWORK_AVAILABLE, _TORCH_LESSER_EQUAL_1_13_1
 
 if _HABANA_FRAMEWORK_AVAILABLE:
@@ -67,7 +66,7 @@ class HPUParallelStrategy(DDPStrategy):
         parallel_devices: Optional[List[torch.device]] = [torch.device("hpu")] * HPUAccelerator.auto_device_count(),
         cluster_environment: Optional[ClusterEnvironment] = None,
         checkpoint_io: Optional[CheckpointIO] = None,
-        precision: Optional[Precision] = None,
+        precision: Optional[HPUPrecision] = None,
         process_group_backend: Optional[str] = "hccl",
         timeout: Optional[timedelta] = default_pg_timeout,
         start_method: Literal["popen", "spawn", "fork", "forkserver"] = "popen",