+++ tests/test_pytorch/test_profiler.py
@@ -34,6 +34,7 @@ elif module_available("pytorch_lightning"):
 
 
 from lightning_habana.pytorch.accelerator import HPUAccelerator
+from lightning_habana.pytorch.plugins.precision import HPUPrecisionPlugin
 from lightning_habana.pytorch.strategies import HPUParallelStrategy, SingleHPUStrategy
 
 if _KINETO_AVAILABLE:
@@ -70,6 +71,7 @@ def test_hpu_simple_profiler_trainer_stages(tmpdir):
     model = BoringModel()
     profiler = SimpleProfiler(dirpath=tmpdir, filename="profiler")
     _strategy = SingleHPUStrategy()
+    _plugins = [HPUPrecisionPlugin(precision="bf16-mixed")]
     trainer = Trainer(
         profiler=profiler,
         accelerator=HPUAccelerator(),